Split playground.ts into focused modules
Extract a PLAYGROUND_HOST constant for the backend URL instead of
hardcoding it in evalSnippet and checkSnippet. Move the Garden
language definition, editor factory, snippet button setup and API
calls into separate files.
