+++ website/static/playground.ts
@@ -1,336 +1,6 @@
-import { EditorView, minimalSetup } from "codemirror";
-import { EditorState } from "@codemirror/state";
-import { HighlightStyle, syntaxHighlighting } from "@codemirror/language";
-import { tags } from "@lezer/highlight";
-import { StreamLanguage } from "@codemirror/language";
-import { history, historyKeymap } from "@codemirror/commands";
-import { keymap, highlightActiveLine } from "@codemirror/view";
-import { defaultKeymap } from "@codemirror/commands";
-import escapeHtml from "escape-html";
-
-// Store EditorView instances for each snippet
-const editorViews = new WeakMap<Element, EditorView>();
-
-// Garden language definition
-const gardenLanguage = StreamLanguage.define({
-  token(stream) {
-    // Comments
-    if (stream.match("//")) {
-      stream.skipToEnd();
-      return "comment";
-    }
-
-    // Strings
-    if (stream.match('"')) {
-      while (!stream.eol()) {
-        if (stream.next() === '"') {
-          break;
-        }
-      }
-      return "string";
-    }
-
-    // Keywords
-    if (
-      stream.match(
-        /\b(as|assert|break|continue|else|enum|for|fun|if|import|in|let|match|method|public|return|shared|struct|test|while)\b/,
-      )
-    ) {
-      return "keyword";
-    }
-
-    // Types (CamelCase words)
-    if (stream.match(/\b[A-Z][a-zA-Z0-9_]*\b/)) {
-      return "type";
-    }
-
-    // Skip other characters
-    stream.next();
-    return null;
-  },
-});
-
-// Syntax highlighting style for Garden
-const gardenHighlighting = syntaxHighlighting(
-  HighlightStyle.define([
-    { tag: tags.comment, color: "gray" },
-    { tag: tags.string, color: "#910a0a" },
-    { tag: tags.keyword, color: "#d05416", fontWeight: "bold" },
-    { tag: tags.typeName, color: "#57269c" },
-  ]),
-);
-
-type StdoutOutput = {
-  printed: {
-    s: string;
-  };
-};
-
-type StderrOutput = {
-  printed_stderr: {
-    s: string;
-  };
-};
-
-type PlaygroundResult = {
-  error: string | null;
-  value: string | null;
-};
-
-type PlaygroundResponse = {
-  success: boolean;
-  results?: (StdoutOutput | StderrOutput | PlaygroundResult)[];
-  error?: string;
-  rawOutput?: string;
-};
-
-type CheckDiagnostic = {
-  line_number: number;
-  end_line_number: number;
-  column: number;
-  end_column: number;
-  message: string;
-  severity: "error" | "warning";
-};
-
-type CheckResponse = {
-  success: boolean;
-  diagnostics?: CheckDiagnostic[];
-  error?: string;
-  rawOutput?: string;
-};
-
-function evalSnippet(src: string, snippetDiv: HTMLElement): void {
-  snippetDiv.hidden = false;
-
-  snippetDiv.innerHTML = '<div class="spinner"></div>';
-
-  // This is the live site. To test local versions:
-  //
-  // $ cd playground
-  // $ npm start
-  //
-  // and change this to localhost:3000.
-  fetch("https://playground.garden-lang.org/run", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ src }),
-  })
-    .then((response) => response.json())
-    .then((data: PlaygroundResponse) => {
-      if (!data.success) {
-        snippetDiv.innerHTML = `Error: ${escapeHtml(data.error || "Unknown error")}`;
-        return;
-      }
-
-      if (!data.results || data.results.length === 0) {
-        snippetDiv.innerHTML = "No output";
-        return;
-      }
-
-      // Iterate over all results, preserving stdout/stderr emission order.
-      const outputParts: string[] = [];
-      const valueParts: string[] = [];
-      let hasError = false;
-
-      for (const item of data.results) {
-        if ("printed" in item) {
-          outputParts.push(escapeHtml(item.printed.s));
-        } else if ("printed_stderr" in item) {
-          outputParts.push(
-            `<span class="stderr">${escapeHtml(item.printed_stderr.s)}</span>`,
-          );
-        } else if ("error" in item || "value" in item) {
-          const result = item;
-
-          if (result.error) {
-            snippetDiv.innerHTML = `Error: ${escapeHtml(result.error)}`;
-            hasError = true;
-            break;
-          }
-
-          if (result.value) {
-            valueParts.push(escapeHtml(result.value));
-          }
-        }
-      }
-
-      if (!hasError) {
-        let output = "";
-        if (outputParts.length > 0) {
-          output = outputParts.join("");
-        }
-        if (valueParts.length > 0) {
-          if (output.length > 0) {
-            output += "\n";
-          }
-          output += valueParts.join("\n");
-        }
-
-        snippetDiv.innerHTML = output;
-      }
-    })
-    .catch((error: unknown) => {
-      const message = error instanceof Error ? error.message : String(error);
-      snippetDiv.innerHTML = `Fetch error: ${escapeHtml(message)}`;
-    });
-}
-
-function checkSnippet(src: string, snippetDiv: HTMLElement): void {
-  snippetDiv.hidden = false;
-
-  snippetDiv.innerHTML = '<div class="spinner"></div>';
-
-  fetch("https://playground.garden-lang.org/check", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({ src }),
-  })
-    .then((response) => response.json())
-    .then((data: CheckResponse) => {
-      if (!data.success) {
-        snippetDiv.innerHTML = `Error: ${escapeHtml(data.error || "Unknown error")}`;
-        return;
-      }
-
-      const diagnostics = data.diagnostics || [];
-      if (diagnostics.length === 0) {
-        snippetDiv.innerHTML = "No issues found.";
-        return;
-      }
-
-      const lines = diagnostics.map((d) => {
-        const label = d.severity === "error" ? "Error" : "Warning";
-        const cls = d.severity === "error" ? "stderr" : "";
-        const prefix = `${label} at line ${d.line_number}, column ${d.column}: `;
-        const text = `${prefix}${d.message}`;
-        return cls
-          ? `<span class="${cls}">${escapeHtml(text)}</span>`
-          : escapeHtml(text);
-      });
-      snippetDiv.innerHTML = lines.join("\n");
-    })
-    .catch((error: unknown) => {
-      const message = error instanceof Error ? error.message : String(error);
-      snippetDiv.innerHTML = `Fetch error: ${escapeHtml(message)}`;
-    });
-}
-
-function resetSnippet(
-  snippetDiv: Element,
-  editorView: EditorView,
-  originalCodeNode: HTMLPreElement | null,
-  editButton: Element,
-): void {
-  const editorDom = editorView.dom;
-
-  if (originalCodeNode) {
-    // Restore original with syntax highlighting
-    editorDom.replaceWith(originalCodeNode);
-    editorView.destroy();
-  }
-
-  editorViews.delete(snippetDiv);
-
-  // Hide the output div
-  const outputDiv = snippetDiv.querySelector(".snippet-output");
-  if (outputDiv instanceof HTMLElement) {
-    outputDiv.hidden = true;
-  }
-
-  // Change button text back to "Edit"
-  editButton.textContent = "Edit";
-}
-
-function setupSnippetButtons(): void {
-  document.querySelectorAll(".snippet").forEach((snippetDiv) => {
-    // Set up run button
-    const runButton = snippetDiv.querySelector(".run-snippet");
-    if (runButton) {
-      runButton.addEventListener("click", (_e) => {
-        let src = "";
-
-        // Check for CodeMirror editor first (edit mode)
-        const editorView = editorViews.get(snippetDiv);
-        if (editorView) {
-          src = editorView.state.sliceDoc();
-        } else {
-          // Check for textarea (legacy edit mode) or pre (view mode)
-          const textarea = snippetDiv.querySelector("textarea");
-          const codeNode = snippetDiv.querySelector("pre");
-
-          if (textarea instanceof HTMLTextAreaElement) {
-            src = textarea.value;
-          } else if (codeNode) {
-            src = codeNode.textContent || "";
-          }
-        }
-
-        const outputDiv = snippetDiv.querySelector(".snippet-output");
-        if (outputDiv instanceof HTMLElement) {
-          evalSnippet(src, outputDiv);
-        }
-      });
-    }
-
-    // Set up edit button
-    const editButton = snippetDiv.querySelector(".edit-snippet");
-    if (editButton) {
-      let originalCodeNode: HTMLPreElement | null = null;
-      let originalTextContent = "";
-
-      editButton.addEventListener("click", (_e) => {
-        const codeNode = snippetDiv.querySelector("pre");
-        const editorView = editorViews.get(snippetDiv);
-
-        // Check if we're in CodeMirror edit mode
-        if (editorView) {
-          // Currently in edit mode, reset to original pre element
-          resetSnippet(snippetDiv, editorView, originalCodeNode, editButton);
-
-          originalCodeNode = null;
-          originalTextContent = "";
-        } else if (codeNode instanceof HTMLPreElement) {
-          // Currently in view mode, switch to CodeMirror edit mode
-          originalCodeNode = codeNode.cloneNode(true) as HTMLPreElement;
-          originalTextContent = codeNode.textContent || "";
-
-          // Create CodeMirror editor
-          const state = EditorState.create({
-            doc: originalTextContent,
-            extensions: [
-              minimalSetup,
-              EditorView.lineWrapping,
-              history(),
-              keymap.of([...defaultKeymap, ...historyKeymap]),
-              highlightActiveLine(),
-              gardenLanguage,
-              gardenHighlighting,
-            ],
-          });
-
-          const newEditorView = new EditorView({
-            state,
-          });
-
-          // Store reference in WeakMap so Run button can access it
-          editorViews.set(snippetDiv, newEditorView);
-
-          codeNode.replaceWith(newEditorView.dom);
-
-          // Change button text to "Reset"
-          editButton.textContent = "Reset";
-
-          newEditorView.focus();
-        }
-      });
-    }
-  });
-}
+import { evalSnippet, checkSnippet } from "./api";
+import { createGardenEditor } from "./editor";
+import { setupSnippetButtons } from "./snippets";
 
 function setupPlayground(): void {
   const playgroundRunButton = document.querySelector("#playground-run");
@@ -346,23 +16,7 @@ function setupPlayground(): void {
   ) {
     const initialContent = playgroundEditor.value + "\n\n";
 
-    // Create CodeMirror editor
-    const state = EditorState.create({
-      doc: initialContent,
-      extensions: [
-        minimalSetup,
-        EditorView.lineWrapping,
-        history(),
-        keymap.of([...defaultKeymap, ...historyKeymap]),
-        highlightActiveLine(),
-        gardenLanguage,
-        gardenHighlighting,
-      ],
-    });
-
-    const editorView = new EditorView({
-      state,
-    });
+    const editorView = createGardenEditor(initialContent);
 
     // Replace textarea with CodeMirror editor
     playgroundEditor.replaceWith(editorView.dom);
